simplify zotero citekey, rename file. see dscr for testing
I can link these notes files to zotero using the markdb-connect plugin IF:
The filename is @citekey
OR there is `citekey: citekay` in the yaml.
The latter requires changing the regex for file search to remove the @. Even when I do that though, it can't identify from the filename alone with the @

THEREFORE: to work with this plugin, either the notes files need to have a yaml tag added, or they need to be renamed to use @ at the start.